add list load HTML class
