+++ javascript/books/bookObj.js
@@ -1,8 +1,7 @@
 import { Book } from './book.js';
+import { loadHTML } from './lists.js';
 
 import { BookList } from './setBookList.js';
-
-const docRange = document.createRange();
 
 const listWrapper = document.getElementById('book-list-wrapper');
 
@@ -32,33 +31,7 @@ class BookStore extends BookList {
 
 const bookStore = new BookStore();
 
-const getHTML = (books) => {
-  let template = '<ul class="book-list">';
-  books.forEach((book, i) => {
-    template += `<li class="list-item ${i % 2 !== 0 ? 'dark' : 'light'}">
-          <div class="book-list-items">
-            <h5 class="book-items book-title">"${book.title}" by ${book.author}</h5>
-            <button id="btn-${i}" data-book-index="${i}" type="button" class="removeBtn">Remove</button>
-          </div>
-        </li>`;
-  });
-
-  if (books.length === 0) {
-    template += `
-      <li class="list-item light">
-        <div class="book-list-items book-list-items-empty">
-          <h5 class="book-items book-title">Book List is Empty</h5>
-        </div>
-      </li>
-    `;
-  }
-
-  template += '</ul>';
-
-  return docRange.createContextualFragment(template);
-};
-
-listWrapper.append(getHTML(bookStore.books));
+listWrapper.append(loadHTML.getHTML(bookStore.books));
 
 export { bookStore };
